refactor: 투표 생성 xml 코드 다듬기
