+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,7 @@
         android:theme="@style/Theme.WooManSi"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.screen.group.VoteCreateActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
+
         <activity
             android:name=".ui.screen.SplashActivity"
             android:exported="true">
@@ -39,6 +33,7 @@
         <activity android:name=".ui.screen.group.GroupInfoActivity" />
         <activity android:name=".ui.screen.group.GroupDetailActivity" />
         <activity android:name=".ui.screen.timetable.createTimeTableActivity" />
+        <activity android:name=".ui.screen.vote.VoteCreateActivity" />
         <activity android:name=".ui.screen.vote.VoteJoinActivity" />
         <activity android:name=".ui.screen.vote.VoteResultActivity" />
     </application>
