+++ app/src/main/res/layout/activity_vote_create_with_appbar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#F3F2F2"
-    tools:context=".ui.screen.group.VoteCreateActivity">
+    tools:context=".ui.screen.vote.VoteCreateActivity">
 
 
     <com.google.android.material.appbar.AppBarLayout
